packages/ih-cdk/pyproject.toml
[project]
name = "ih-cdk"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aws-cdk-lib>=2.224.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
